backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2
pytest==7.4.0